fix: price 자료형 int -> double

+++ src/main/java/com/example/kiosk_1/MenuItem.java
@@ -9,11 +9,11 @@ package com.example.kiosk_1;
 public class MenuItem {
     // Class Attribute : 음식 이름, 가격, 설명
     private String itemName;
-    private int itemPrice;
+    private double itemPrice;
     private String itemDescription;
 
     // 생성자
-    public MenuItem(String itemName, int itemPrice, String itemDescription) {
+    public MenuItem(String itemName, double itemPrice, String itemDescription) {
         this.itemName = itemName;
         this.itemPrice = itemPrice;
         this.itemDescription = itemDescription;
@@ -21,10 +21,10 @@ public class MenuItem {
 
     // Getter, Setter 메서드를 활용한 Capsulation
     public String getName(){return itemName;}
-    public int getPrice(){ return itemPrice;}
+    public double getPrice(){ return itemPrice;}
     public String getDescription(){return itemDescription;}
     public boolean setName(String name) { itemName = name; return true; }
-    public boolean setPrice(int price) { itemPrice = price; return true;}
+    public boolean setPrice(double price) { itemPrice = price; return true;}
     public boolean setDescription(String dsc) { itemDescription = dsc; return true;}
 
 }
